pyproject.toml
[project]
name = "passive-income-generator"
version = "1.0.0"
description = "Automated digital product generator for Etsy passive income"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "fpdf2>=2.7",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "httpx>=0.25",
]

[project.scripts]
generate = "src.main:main"
